+++ src/main/java/com/amidos/EmailDemo/SendEmail/SendEmail.java
@@ -1,3 +1,4 @@
+package com.amidos.EmailDemo.SendEmail;
 import java.util.*;
 import javax.mail.*;
 import javax.mail.internet.*;
